+++ configurations/sso.template.ini.php
@@ -1,1 +1,0 @@
-login_link = @LOGIN_LINK_SSO@