+++ admin_console/configs/lang/en.php
@@ -1514,8 +1514,12 @@ return array(
 	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::FI' => "Finnish",
 	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::IS' => "Icelandic",
 	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::PL' => "Polish",
-	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::ID' => "Indonesian",
 	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::IN' => "Indonesian",
 	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::EL' => "Greek",
 	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::RO' => "Romanian",
+	'Kaltura_Client_Reach_Enum_CatalogItemLanguage::HU' => "Hungarian",
+
+	'Kaltura_Client_Sso_Enum_SsoStatus::DISABLED' => 'Disabled',
+	'Kaltura_Client_Sso_Enum_SsoStatus::ACTIVE' => 'Enabled',
+	'Kaltura_Client_Sso_Enum_SsoStatus::DELETED' => 'Deleted',
 );