+++ deployment/updates/scripts/add_permissions/2019_09_11_add_webcasting_role_permission_to_user_list.php
@@ -1,0 +1,8 @@
+<?php
+/**
+ * @package deployment
+ * @subpackage orion.roles_and_permissions
+ */
+$addPermissionsScript = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$addConfig = realpath(dirname(__FILE__)) . '/../../../permissions/service.user.ini';
+passthru("php $addPermissionsScript $addConfig");