+++ release-notes.md
@@ -5,7 +5,7 @@
 - Issue ID: PLAT-9973
 
 ### configuration ###
-To use the sso emails you should set the sso login link in sso.template.ini
+To use the sso emails you should set the sso login link in sso.ini
 
 ## Adding flavorparam permission to Capture app ##
 - Issue Type: Task